parse response MIME, check the charset and lang
If the charset is not UTF-8 nor ASCII then we don't know how to handle
the body and so rejects it.  An abset charset is implied to be utf8,
so we're good.

In the generated HTML we always use UTF8 charset, which is fine too,
since it is a superset of ASCII.

Attempt also at saving the lang for later reuse in the generated HTML.